# swagger-api.json
{"openapi":"3.0.0","info":{"title":"ClÃ­nicaCRM API","description":"API for managing ophthalmology clinic cases and patient leads. This API supports both session-based authentication (for browser requests) and API key authentication (for programmatic access).","version":"1.0.0","contact":{"name":"ClÃ­nicaCRM Support","email":"support@clinicacrm.com"}},"servers":[{"url":"http://localhost:3000/api","description":"Development server"}],"paths":{"/api-keys":{"get":{"summary":"Get user API keys","description":"Retrieve all API keys for the authenticated user","security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApiKey"}}}}},"401":{"description":"Unauthorized - User not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}}}},"post":{"summary":"Create new API key","description":"Generate a new API key for the authenticated user","security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string","description":"Name for the API key"},"description":{"type":"string","description":"Optional description"},"expiresInDays":{"type":"number","description":"Number of days until expiration (optional)"},"permissions":{"type":"array","items":{"type":"string","enum":["read","write","admin"]},"default":["read","write"],"description":"Permissions for the API key"},"keyType":{"type":"string","enum":["secure","readable","prefixed"],"default":"secure","description":"Type of API key to generate"}}}}}},"responses":{"200":{"description":"API key created successfully","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"key":{"type":"string","description":"The generated API key (only shown once)"},"name":{"type":"string"},"description":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"expires_at":{"type":"string","format":"date-time"},"permissions":{"type":"array","items":{"type":"string"}},"message":{"type":"string","example":"API key created successfully. Please save this key securely - it will not be shown again."}}}}}},"400":{"description":"Bad request - Missing required fields","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Name is required"}}}}}},"401":{"description":"Unauthorized - User not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}}}},"put":{"summary":"Update API key","description":"Update an existing API key","security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id"],"properties":{"id":{"type":"string","format":"uuid","description":"API key ID"},"name":{"type":"string","description":"New name for the API key"},"description":{"type":"string","description":"New description"},"is_active":{"type":"boolean","description":"Whether the API key is active"},"permissions":{"type":"array","items":{"type":"string","enum":["read","write","admin"]},"description":"New permissions for the API key"},"expiresInDays":{"type":"number","description":"Number of days until expiration (null to remove expiration)"}}}}}},"responses":{"200":{"description":"API key updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiKey"}}}},"400":{"description":"Bad request - Missing API key ID","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"API key ID is required"}}}}}},"401":{"description":"Unauthorized - User not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}},"404":{"description":"API key not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"API key not found"}}}}}}}},"delete":{"summary":"Delete API key","description":"Delete an API key by ID","security":[{"BearerAuth":[]}],"parameters":[{"name":"id","in":"query","required":true,"description":"API key ID to delete","schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"API key deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"API key deleted successfully"}}}}}},"400":{"description":"Bad request - Missing API key ID","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"API key ID is required"}}}}}},"401":{"description":"Unauthorized - User not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}}}}},"/lookup":{"get":{"summary":"Lookup case by phone number for CCaaS screen pop","description":"Search for a case using a phone number. Returns case ID and essential data optimized for CCaaS screen pop functionality. Searches across all phone fields (phone, home_phone, cell_phone).","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"parameters":[{"name":"phone","in":"query","required":true,"description":"Phone number to search for (e.g., +391234567890)","schema":{"type":"string","example":"+391234567890"}}],"responses":{"200":{"description":"Case found successfully","content":{"application/json":{"schema":{"type":"object","properties":{"found":{"type":"boolean","example":true},"case_id":{"type":"string","format":"uuid","description":"The case ID for screen pop"},"patient":{"type":"object","properties":{"name":{"type":"string","example":"Mario Rossi"},"phone":{"type":"string","example":"+391234567890"},"home_phone":{"type":"string","example":"+391234567891"},"cell_phone":{"type":"string","example":"+391234567892"},"email":{"type":"string","example":"mario.rossi@email.com"}}},"case":{"type":"object","properties":{"status":{"type":"string","example":"APPUNTAMENTO"},"clinic":{"type":"string","example":"Clinica Oculistica Roma"},"treatment":{"type":"string","example":"Cataratta"},"disposition":{"type":"string","example":"Interessato"},"outcome":{"type":"string","example":"Paziente conferma appuntamento"},"created_at":{"type":"string","format":"date-time"},"follow_up_date":{"type":"string","format":"date-time"}}},"screen_pop_url":{"type":"string","description":"Direct URL to open the case in the CRM","example":"https://crm-gules-six.vercel.app/it/cases/123e4567-e89b-12d3-a456-426614174000"},"phone_searched":{"type":"string","example":"+391234567890"}}}}}},"400":{"description":"Bad request - Missing phone number","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Phone number is required"},"message":{"type":"string","example":"Please provide a phone number in the query parameter: ?phone=+391234567890"}}}}}},"401":{"description":"Unauthorized - Invalid or missing API key","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}},"404":{"description":"No case found for the phone number","content":{"application/json":{"schema":{"type":"object","properties":{"found":{"type":"boolean","example":false},"message":{"type":"string","example":"No case found for phone number: +391234567890"},"phone":{"type":"string","example":"+391234567890"}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"post":{"summary":"Lookup case by phone number (POST method)","description":"Alternative POST method for phone lookup. Send phone number in request body instead of query parameter.","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["phone"],"properties":{"phone":{"type":"string","description":"Phone number to search for","example":"+391234567890"}}}}}},"responses":{"200":{"description":"Case found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LookupResponse"}}}},"400":{"description":"Bad request - Missing phone number","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Phone number is required in request body"},"message":{"type":"string","example":"Please provide a phone number in the request body: {\"phone\": \"+391234567890\"}"}}}}}},"401":{"description":"Unauthorized - Invalid or missing API key","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"}}}}}},"404":{"description":"No case found for the phone number","content":{"application/json":{"schema":{"type":"object","properties":{"found":{"type":"boolean","example":false},"message":{"type":"string","example":"No case found for phone number: +391234567890"},"phone":{"type":"string","example":"+391234567890"}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/cases":{"get":{"summary":"Get all cases","description":"Retrieve all cases with optional filtering","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"parameters":[{"name":"search","in":"query","description":"Search by first name, last name, email, or phone","schema":{"type":"string"}},{"name":"status","in":"query","description":"Filter by case status","schema":{"type":"string","enum":["IN_CORSO","APPUNTAMENTO","CHIUSO"]}},{"name":"channel","in":"query","description":"Filter by lead channel","schema":{"type":"string","enum":["WEB","FACEBOOK","INFLUENCER"]}},{"name":"date_from","in":"query","description":"Filter cases created from this date (ISO format)","schema":{"type":"string","format":"date-time"}},{"name":"date_to","in":"query","description":"Filter cases created until this date (ISO format)","schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Case"}}}}},"401":{"description":"Unauthorized - Invalid or missing API key","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Unauthorized"},"message":{"type":"string","example":"Invalid or missing API key. Please provide a valid x-api-key header or Bearer token."}}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}},"post":{"summary":"Create a new case","description":"Create a new patient case/lead","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["first_name","last_name","phone","channel","origin","clinic"],"properties":{"assigned_to":{"type":"string","format":"uuid","description":"User ID assigned to this case"},"first_name":{"type":"string","description":"Patient first name"},"last_name":{"type":"string","description":"Patient last name"},"phone":{"type":"string","description":"Patient phone number"},"email":{"type":"string","format":"email","description":"Patient email address"},"channel":{"type":"string","enum":["WEB","FACEBOOK","INFLUENCER"],"description":"Lead source channel"},"origin":{"type":"string","description":"Specific origin of the lead"},"status":{"type":"string","enum":["IN_CORSO","APPUNTAMENTO","CHIUSO"],"default":"IN_CORSO","description":"Case status"},"outcome":{"type":"string","description":"Current outcome or notes"},"clinic":{"type":"string","description":"Clinic name"},"treatment":{"type":"string","description":"Treatment type"},"promotion":{"type":"string","description":"Promotion or campaign"},"follow_up_date":{"type":"string","format":"date-time","description":"Follow-up appointment date"},"dialer_campaign_tag":{"type":"string","description":"Campaign tag for dialer integration"}}}}}},"responses":{"200":{"description":"Case created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Case"}}}},"400":{"description":"Bad request - validation error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"500":{"description":"Internal server error"}}},"put":{"summary":"Update a case (full update)","description":"Update all fields of an existing case","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id"],"properties":{"id":{"type":"integer","description":"Case ID to update"},"assigned_to":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string","format":"email"},"channel":{"type":"string","enum":["WEB","FACEBOOK","INFLUENCER"]},"origin":{"type":"string"},"status":{"type":"string","enum":["IN_CORSO","APPUNTAMENTO","CHIUSO"]},"outcome":{"type":"string"},"clinic":{"type":"string"},"treatment":{"type":"string"},"promotion":{"type":"string"},"follow_up_date":{"type":"string","format":"date-time"},"dialer_campaign_tag":{"type":"string"}}}}}},"responses":{"200":{"description":"Case updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Case"}}}},"400":{"description":"Bad request - missing case ID"},"404":{"description":"Case not found"},"500":{"description":"Internal server error"}}},"patch":{"summary":"Update a case (partial update)","description":"Update specific fields of an existing case","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id"],"properties":{"id":{"type":"integer","description":"Case ID to update"},"assigned_to":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string","format":"email"},"channel":{"type":"string","enum":["WEB","FACEBOOK","INFLUENCER"]},"origin":{"type":"string"},"status":{"type":"string","enum":["IN_CORSO","APPUNTAMENTO","CHIUSO"]},"outcome":{"type":"string"},"clinic":{"type":"string"},"treatment":{"type":"string"},"promotion":{"type":"string"},"follow_up_date":{"type":"string","format":"date-time"},"dialer_campaign_tag":{"type":"string"}}}}}},"responses":{"200":{"description":"Case updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Case"}}}},"400":{"description":"Bad request - missing case ID"},"404":{"description":"Case not found"},"500":{"description":"Internal server error"}}}},"/cases/delete":{"delete":{"summary":"Delete a case","description":"Delete a case by ID","security":[{"ApiKeyAuth":[]},{"BearerAuth":[]}],"parameters":[{"name":"id","in":"query","required":true,"description":"Case ID to delete","schema":{"type":"integer"}}],"responses":{"200":{"description":"Case deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}}}}}},"400":{"description":"Bad request - missing case ID"},"500":{"description":"Internal server error"}}}}},"components":{"securitySchemes":{"ApiKeyAuth":{"type":"apiKey","in":"header","name":"x-api-key","description":"API key for programmatic access"},"BearerAuth":{"type":"http","scheme":"bearer","description":"Bearer token (alternative to x-api-key header)"}},"schemas":{"ApiKey":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique API key identifier"},"name":{"type":"string","description":"API key name"},"description":{"type":"string","description":"Optional description"},"created_at":{"type":"string","format":"date-time","description":"Creation timestamp"},"expires_at":{"type":"string","format":"date-time","description":"Expiration timestamp (optional)"},"is_active":{"type":"boolean","description":"Whether the API key is active"},"permissions":{"type":"array","items":{"type":"string","enum":["read","write","admin"]},"description":"API key permissions"},"last_used":{"type":"string","format":"date-time","description":"Last usage timestamp"},"usage_count":{"type":"integer","description":"Number of times the key has been used"},"key_preview":{"type":"string","description":"Preview of the key hash (first and last few characters)"}},"required":["id","name","created_at","is_active","permissions","usage_count","key_preview"]},"Case":{"type":"object","properties":{"id":{"type":"integer","description":"Unique case identifier"},"created_at":{"type":"string","format":"date-time","description":"Case creation timestamp"},"assigned_to":{"type":"string","format":"uuid","description":"User ID assigned to this case"},"first_name":{"type":"string","description":"Patient first name"},"last_name":{"type":"string","description":"Patient last name"},"phone":{"type":"string","description":"Patient phone number"},"email":{"type":"string","format":"email","description":"Patient email address"},"channel":{"type":"string","enum":["WEB","FACEBOOK","INFLUENCER"],"description":"Lead source channel"},"origin":{"type":"string","description":"Specific origin of the lead"},"status":{"type":"string","enum":["IN_CORSO","APPUNTAMENTO","CHIUSO"],"description":"Case status"},"outcome":{"type":"string","description":"Current outcome or notes"},"clinic":{"type":"string","description":"Clinic name"},"treatment":{"type":"string","description":"Treatment type"},"promotion":{"type":"string","description":"Promotion or campaign"},"follow_up_date":{"type":"string","format":"date-time","description":"Follow-up appointment date"},"dialer_campaign_tag":{"type":"string","description":"Campaign tag for dialer integration"}},"required":["id","created_at","first_name","last_name","phone","channel","origin","clinic"]},"LookupResponse":{"type":"object","properties":{"found":{"type":"boolean","description":"Whether a case was found for the phone number"},"case_id":{"type":"string","format":"uuid","description":"The case ID for screen pop"},"patient":{"type":"object","properties":{"name":{"type":"string","description":"Full patient name"},"phone":{"type":"string","description":"Primary phone number"},"home_phone":{"type":"string","description":"Home phone number"},"cell_phone":{"type":"string","description":"Cell phone number"},"email":{"type":"string","description":"Patient email address"}}},"case":{"type":"object","properties":{"status":{"type":"string","description":"Current case status"},"clinic":{"type":"string","description":"Clinic name"},"treatment":{"type":"string","description":"Treatment type"},"disposition":{"type":"string","description":"Patient disposition"},"outcome":{"type":"string","description":"Current outcome"},"created_at":{"type":"string","format":"date-time","description":"Case creation date"},"follow_up_date":{"type":"string","format":"date-time","description":"Follow-up appointment date"}}},"screen_pop_url":{"type":"string","description":"Direct URL to open the case in the CRM"},"phone_searched":{"type":"string","description":"The phone number that was searched"}},"required":["found","phone_searched"]}}}}